feat: Implement SEO metadata, refactor pages into client/server components, and configure domain redirects and sitemap.
