+++ app/robots.ts
@@ -2,7 +2,7 @@ import { MetadataRoute } from "next";
 
 export default function robots(): MetadataRoute.Robots {
   return {
-    sitemap: "https://handiani.my.id/sitemap.xml",
+    sitemap: "https://gaweqr.my.id/sitemap.xml",
     rules: [
       {
         userAgent: ["Googlebot", "Applebot", "Bingbot"],
