+++ app/templates/page.tsx
@@ -1,105 +1,30 @@
-"use client";
+import type { Metadata } from "next";
+import QRTemplates from "@/app/templates/page-client";
 
-import Icon from "@/components/Shared/Icon";
-import { templates } from "@/constants/template.data";
-import { motion } from "framer-motion";
-import { ArrowRight, Sparkles } from "lucide-react";
-import Link from "next/link";
-
-const container = {
-  hidden: { opacity: 0 },
-  show: {
-    opacity: 1,
-    transition: { staggerChildren: 0.08, delayChildren: 0.1 },
+export const metadata: Metadata = {
+  title: "QR Code Templates",
+  description:
+    "Explore ready-to-use QR templates for links, WiFi, WhatsApp, contacts, and email. Start faster with GaweQR templates.",
+  alternates: {
+    canonical: "/templates",
+  },
+  openGraph: {
+    title: "QR Code Templates - GaweQR",
+    description:
+      "Explore ready-to-use QR templates for links, WiFi, WhatsApp, contacts, and email. Start faster with GaweQR templates.",
+    url: "/templates",
+    images: ["/qr_code.png"],
+    type: "website",
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "QR Code Templates - GaweQR",
+    description:
+      "Explore ready-to-use QR templates for links, WiFi, WhatsApp, contacts, and email. Start faster with GaweQR templates.",
+    images: ["/qr_code.png"],
   },
 };
 
-const item = {
-  hidden: { opacity: 0, y: 18 },
-  show: { opacity: 1, y: 0, transition: { duration: 0.45, ease: "easeOut" } },
-};
-
 export default function TemplatesPage() {
-  return (
-    <main className="relative min-h-[calc(100dvh-4rem)] overflow-hidden px-4 pb-16 pt-12 sm:px-6 lg:px-8">
-      <div className="pointer-events-none absolute inset-0 -z-10">
-        <div className="absolute -top-24 left-[-8%] h-72 w-72 rounded-full bg-primary/15 blur-3xl" />
-        <div className="absolute right-[-6%] top-12 h-80 w-80 rounded-full bg-sky-400/15 blur-3xl" />
-        <div className="absolute bottom-0 left-1/2 h-72 w-72 -translate-x-1/2 rounded-full bg-emerald-300/10 blur-3xl" />
-      </div>
-
-      <motion.section
-        variants={container}
-        initial="hidden"
-        animate="show"
-        className="mx-auto w-full max-w-7xl"
-      >
-        <motion.div variants={item} className="mb-8 text-center sm:mb-10">
-          <p className="inline-flex items-center gap-2 rounded-full border border-primary/20 bg-primary/10 px-3 py-1 text-xs font-medium text-primary">
-            <Sparkles size={14} />
-            Choose Your Starting Point
-          </p>
-          <h1 className="mt-3 text-3xl font-semibold tracking-tight sm:text-4xl lg:text-5xl">
-            Templates built for real workflows
-          </h1>
-          <p className="mx-auto mt-3 max-w-2xl text-sm text-foreground/75 sm:text-base">
-            Start from a purpose-built template, customize style in Studio, and export in seconds.
-          </p>
-        </motion.div>
-
-        <div className="grid gap-4 sm:grid-cols-2 xl:gap-5">
-          {templates.map((template) => {
-            return (
-              <motion.article key={template.title} variants={item}>
-                <Link
-                  href={template.href}
-                  className="group relative block overflow-hidden rounded-3xl border border-foreground/10 bg-background/70 p-5 shadow-sm backdrop-blur transition-all duration-300 hover:-translate-y-1 hover:border-primary/30 hover:shadow-xl hover:shadow-primary/10 sm:p-6"
-                >
-                  <div
-                    className={`pointer-events-none absolute inset-0 bg-gradient-to-br ${template.accent}`}
-                  />
-
-                  <div className="relative space-y-4">
-                    <div className="flex items-start justify-between gap-3">
-                      <div>
-                        <p className="text-xs font-medium uppercase tracking-wide text-primary/90">
-                          {template.subtitle}
-                        </p>
-                        <h2 className="text-xl font-semibold">{template.title}</h2>
-                      </div>
-
-                      <span className="inline-flex h-10 w-10 items-center justify-center rounded-xl border border-foreground/10 bg-background/80 text-primary">
-                        <Icon name={template.icon} size={18} />
-                      </span>
-                    </div>
-
-                    <p className="text-sm text-foreground/75">{template.description}</p>
-
-                    <div className="flex flex-wrap gap-2">
-                      {template.highlights.map((highlight) => (
-                        <span
-                          key={highlight}
-                          className="rounded-full border border-foreground/10 bg-background/80 px-2.5 py-1 text-xs text-foreground/80"
-                        >
-                          {highlight}
-                        </span>
-                      ))}
-                    </div>
-
-                    <span className="inline-flex items-center gap-2 text-sm font-medium text-primary">
-                      Open in Studio
-                      <ArrowRight
-                        size={14}
-                        className="transition-transform duration-200 group-hover:translate-x-1"
-                      />
-                    </span>
-                  </div>
-                </Link>
-              </motion.article>
-            );
-          })}
-        </div>
-      </motion.section>
-    </main>
-  );
+  return <QRTemplates />;
 }
