+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "qr-generator",
+  "name": "gaweqr",
   "version": "0.1.0",
   "private": true,
   "scripts": {
